User-agent header including version storage change
Summary:
Similar to D1065 but for Ruby. User-agent now contains SDK version.

I used an almost identical method, ie storing the version in its own file and importing it when building the gem. A side effect is that this appears to break the `rake version:bump` task, and versions will need to be edited in `lib/version.py` in future.

Test Plan: tested locally to check headers, built gem locally and it works with new versioning.

Reviewers: khamidou

Reviewed By: khamidou

Differential Revision: https://review.inboxapp.com/D1070