[Ruby] Fix issue #7 on github
Summary:
Executive summary: the destroy method wasn't working and people are complaining about it on github (https://github.com/inboxapp/inbox-ruby/issues/7)

We were using a block to handle the response returned by `RestClient`. The problem was that in the case of a `DELETE`,  `RestClient` passes to the block a different number of arguments than for other HTTP requests, which caused an error.

Test Plan: Tested manually + wrote regression test.

Reviewers: jennie

Reviewed By: jennie

Projects: #ruby-sdk

Differential Revision: https://review.inboxapp.com/D1085