[Ruby] Fix shared pointer bug
Summary:
This bug is both serious and funny ;) It's the reason why Jennie wasn't able to use my "Steam data" webapp even though the code seemed correct.

In my steam data app, I have the following innocuous line of code:
```
 steam_messages = namespace.messages.where(:from => 'service@paypal.fr').all +
                   namespace.messages.where(:from => 'service@paypal.com').all +
                   namespace.messages.where(:from => 'service@paypal.co.uk').all +
                   namespace.messages.where(:from => 'service@paypal.co.uk').all +
                   namespace.messages.where(:subject => 'Thank you for your purchase').all +
                   namespace.messages.where(:subject => 'Thank you for your purchase!').all
```
Basically, we're getting messages with various from/subject headers. The problem is that the implementation of `where`, inside the Ruby SDK, does this:

`collection.filters ||= {}`

So the filters dictionary is shared among all instances of RestfulModelCollection. I changed the code to do a deepcopy instead, which fixed the problem (note that Ruby doesn't seem to have a built-in way to do such a thing so we're marshaling/unmarshaling).

Test Plan: Tested manually.

Reviewers: jennie, bengotow

Reviewed By: jennie, bengotow

Projects: #ruby-sdk

Differential Revision: https://phab.nylas.com/D1629