Fix the Ruby SDK for open source users
Summary:
Here's the problem. Because of @mg's latest changes to the API, we now require opensource user to pass an auth token -- the account id -- to access the API.

Unfortunately, there's no way currently for an opensource user of the SDK to get the account id without already being logged in. I fixed this by adding a new resource, `opensource.accounts`, which can be accessed without being logged in.

Test Plan: Tested manually + wrote a tiny test.

Reviewers: mg, bengotow

Reviewed By: bengotow

Subscribers: mg

Projects: #sdks, #api

Differential Revision: https://phab.nylas.com/D1920