Allow limit and offset filters for collections
The limit and offset filters, which were set correctly in the `where`
function, were being overriden in the `get_model_collection` function
by the offset and limit parameters passed to it (with defaults). The
function now checks if limit or offset are already present in the list
of filters before applying its own.

Close issue #66.