Refactor limit and offset to set in filters
Previously, limit and offset could be specified in `@filters` (which
were usually set inside a higher-level function such as `where` or
`range`) and also as function parameters to the lower-level function
`get_model_collection`. The decision as to whether or not the passed
parameters would override those already present in `filters` was
confusing and has been removed in favor of only using the `@filters`
method.