Remove implicit dependency on ActiveSupport
For some reason, we had a require for `active_support` in lib/nylas,
which meant that even though we didn't declare a dependency on
ActiveSupport, we assumed that you would be using it.

This removes that in favor of a small class that casts hashes to query
strings.

We may be able to get rid of this if we leverage RestClient for
URL building.