Use YARD to build docs for better linking in docs (#154)
YARD is the current standard for documentation in Ruby. It provides a similar interface as most other language libraries, including the ability to define the types of params and return values.

Bringing it in will make life a little bit easier as we add more detailed method level documentation through the course of the 4.X release.