Fix exception raise when content-type is not JSON (#303)
Background:
Currently if server returns any error where content-type is not a JSON
is not raising error in `Nylas` namespace and returns
`RestClient::Response` instead. This issue is also raised in #274
and #272. It has been found that it breaks after #268 where we 
have fixed parsing error if content type is not json.

This PR makes sure that whatever content-type is returned from
API in error response it always raises error in `Nylas`
namespace.

Changes:
- Update `Nylas::HttpClient` to parse error response beforing
  raising exception.