[61539] Draft version not updating after `update` (#304)
When making updates to the draft object on the server side (via `.update`) the version number is incremented upon successful update. However, the local object does not update the value when it's returned from the server. This causes an issue when, the user attempts to re-update the draft a second time and the version number has not been incremented. Now, when calling `.update` the version is updated with the value returned from the payload.

Co-authored-by: Arun Agrawal <arunagw@gmail.com>