[63670] Fix `api` value resetting after `message.expanded` call (#311)
When calling message.expanded, the attributes of the message that have an api value (such as files) do not retain the API value after the object's values were updated. The reason for this is that the transfer_attributes method is never called, thus the message object never passes the api value to the newly parsed data. Now we manually invoke the method call to ensure all of the attributes that need an api value get initialized with it.