[64639] Return API error regardless if response is JSON (#316)
In the rare case, the API server might return a server error but with a HTML body instead of a JSON type. This causes the parser to throw a parse error, but it does not return that the server ran into an error. Now, if parsing fails we also check if we encountered an error and if we did, we should propagate the error back to the user for better information.