[67394] Fix updating `Draft` failing if no version number provided (#323)
The following is the code snippet is found on our Nylas docs for Ruby PUT /drafts/{id}: https://developer.nylas.com/docs/api/#put/drafts/id

However in practice this fails. The reason is because the API expects a version number to be passed in with a PUT call. In the Ruby SDK we only pass the updated parameters to the API, so if the user doesn't explicitly send in the version number, the SDK doesn't do it for them and triggers the error. To adhere to the documentation and the API's expectation, we should send the version number in the payload if the user never adds it in explicitly.