[68888] Modify Draft Sending logic (#325)
We shouldn't be updating a draft every time the user tries to to send a draft. This causes an unnecessary PUT call, and potentially causes a delay in the syncing process for the end user. So now, calling .send on a draft will not trigger an update/PUT call. Also to make up for this change, if a user wants to send a draft that they haven't saved (have not triggered a POST /drafts call), the SDK will send the full body in the POST call to /send allowing the user to send a message directly without having to create and save a draft object first.