[65715] Add Event conferencing support (#320)
This PR enables support for the new Nylas Conference Sync (Beta) feature as well as support for auto generating conference details. If you have an integration authorized and enabled for your Nylas account you can provide an autocreate object instead of a details object to have the Nylas API autofill the conferencing details.