Update draft send method to send existing draft when tracking is enabled (#327)
This fixes a bug where if tracking information is enabled on a draft it sends directly rather than sending the existing draft. If a draft has already been saved this means the actual draft message never gets sent and a new message is created. The message should instead only be sent directly if it hasn't been saved and otherwise conditionally add the tracking information to the draft message.