[69622] Allow `exchange_code_for_token` to return a full body (#336)
The exchange_code_for_token method strips the response body from the server and returns just the access_token. Now, it allows the user to pass in a boolean to dictate whether or not to return the full server response. To preserve backwards compatibility, the default behaviour of the method will still return an access_token.