[75450] Fix bug where updating an Event results in an API error (#340)
A recent API update caused updating an event using the SDK to return an error because the SDK sends a Participant's status as well as an object key within the Event.when object. This PR also modifies how we serialize child objects for API calls, and ensure if we don't want to include read_only values that we enforce it throughout the child objects.