Update interval keyword arg for calendar availability methods
This commit updates the keyword arguments for the interval parameter for
the availability-related methods to match the requirements of the API
endpoints.

According to the latest API, the availability endpoints expect an
`interval_minutes` rather than an `interval` keyword. Without this
change, the API throws an error:

```ruby
response = nylas_client.calendars.availability(
  start_time: start_time.to_i,
  end_time: end_time.to_i,
  duration_minutes: duration_minutes,
  interval: 30,
  round_robin: "max-availability",
  emails: [coach.email],
)

=> Nylas::InvalidRequest: interval_minutes: Missing data for required field.
```