Support calendars keyword for availability requests (#350)
This commit adds an optional keyword argument `calendar` to the
Nylas::CalendarCollection's availability related methods. This is
parameter is accepted by the latest API, but was not previously included
in the method definitions.

The argument requirements match the parameter requirements of the
endpoint. A user must provide on of `calendar` or `emails`

Co-authored-by: Anthony M <anthony.moffa@thoughtbot.com>