Fix calendar availability failing when `round_robin` is `nil` (#375)
This PR address the issue where, if round_robin is left as nil, it always yields an error. This is because the Nylas Docs were out of date for the availability endpoint. round_robin is optional, and when omitted, returns collective availability. However, sending null is not valid despite what it says in the documentation. Now, we omit the value from the json payload if the value is nil.