Add support for getting `ids` for collections not supported by the API (#380)
This PR enables the ability for getting a list of ids for every collection, regardless of their support on the API side. We do this by introducing a new collection operation value called id_listable which indicates if the API supports a list view or not. If it does not, we just perform a general GET call and return a array of id strings.