Issue-394: Add disable_provider_selection kwarg to Api.authentication_url
We need the option to disable the "Select different provider" button
in Nylas' hosted auth when we are trying to authenticate against a
specific provider. This is necessary to prevent our users from escaping
the intended auth flow and selecting a provider that we do not support.

This commit addresses this need by adding and optional keyword argument
to the Api.authentication_url method and, if it is provided, including it
in the params payload appended on the call to the /oauth/authorize endpoint.
Tests have been added to account for this change and I tested it locally to
confirm that the "Select different provider" button was no longer visible
when I included disable_provider_selection: true in my call to
the authentication_url method.